papers/CoLaDa/requirements.txt
numpy
torch
seqeval
scipy
faiss-cpu
transformers